% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST.mtx.R
\name{ST.mtx}
\alias{ST.mtx}
\title{Creating Seurat object from mtx}
\usage{
ST.mtx(path, name)
}
\arguments{
\item{Path}{Path to the SpaceRnager output directory which contains the filtered_feature_bc_matrix folder and spatial folders}

\item{Name}{Name of the project}
}
\value{
A Visium Seurat object
}
\description{
Normally the Visium Seurat objects are made directly from HDF5 feature-barcade matrix format.
However, somethimes the HDF5 (.h5) format is not available when using public data.
Using ST.mtx, one can create the Visium Seurat objetcs directly from the gene
count matrix and Spatial transcriptomics images.
}
